Added hasOne Relationship
Since we are using through model to persist relations in out db.
Therefore we need to always create relations using through model.
But in case we need a hasOne relationship, we can first create a
relationship as we could with a through model, and then add this to the
model.json

"mixins": {
"HasOne": {
"relations": [
{
"model": "<MODEL_NAME>",
"hasManyRelation": "<NAME_OF_THROUGH_MODEL>"
}
]
}
}

it will expose the hasOne api with the <MODEL_NAME>.

You can add multiple relations in array if needed.